+++ src/desktop-affine.h
@@ -28,14 +28,14 @@ NRMatrix *sp_desktop_doc2w_affine (SPDesktop *desktop, NRMatrix *doc2w);
 NRMatrix *sp_desktop_root2dt_affine (SPDesktop *dt, NRMatrix *root2dt);
 NRMatrix *sp_desktop_dt2root_affine (SPDesktop *dt, NRMatrix *dt2root);
 
-NRPoint *sp_desktop_w2d_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_d2w_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_w2doc_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_doc2w_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_d2doc_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_doc2d_xy_point (SPDesktop *desktop, NRPoint *p, float x, float y);
+NRPoint *sp_desktop_w2d_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_d2w_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_w2doc_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_doc2w_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_d2doc_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_doc2d_xy_point (SPDesktop *desktop, NRPoint *p, double x, double y);
 
-NRPoint *sp_desktop_root2dt_xy_point (SPDesktop *dt, NRPoint *p, float x, float y);
-NRPoint *sp_desktop_dt2root_xy_point (SPDesktop *dt, NRPoint *p, float x, float y);
+NRPoint *sp_desktop_root2dt_xy_point (SPDesktop *dt, NRPoint *p, double x, double y);
+NRPoint *sp_desktop_dt2root_xy_point (SPDesktop *dt, NRPoint *p, double x, double y);
 
 #endif
